Identify legacy divisions in search options
Jira ticket: CAMS-473
