+++ user-interface/src/data-verification/dataVerificationHelper.test.ts
@@ -1,6 +1,6 @@
 import { describe } from 'vitest';
 import { CourtDivisionDetails } from '@common/cams/courts';
-import { getOfficeList } from './dataVerificationHelper';
+import { courtSorter, getOfficeList } from './dataVerificationHelper';
 
 describe('data verification helper tests', () => {
   test('should properly map court information for selection', () => {
@@ -94,5 +94,135 @@ describe('data verification helper tests', () => {
 
     const actualOptions = getOfficeList(sortedTestOffices);
     expect(actualOptions).toStrictEqual(expectedOptions);
+  });
+
+  test('should label legacy offices in the options', () => {
+    const testOffices: CourtDivisionDetails[] = [
+      {
+        courtDivisionCode: '002',
+        groupDesignator: 'AB',
+        courtId: '0102',
+        officeCode: '2',
+        officeName: 'B1',
+        state: 'NY',
+        courtName: 'B',
+        courtDivisionName: 'New York 1',
+        regionId: '02',
+        regionName: 'NEW YORK',
+        isLegacy: true,
+      },
+    ];
+
+    const expectedOptions: Array<Record<string, string>> = [
+      { value: '002', label: 'B (New York 1) Legacy' },
+    ];
+
+    const actualOptions = getOfficeList(testOffices);
+    expect(actualOptions).toStrictEqual(expectedOptions);
+  });
+
+  test('should sort offices', () => {
+    const testOffices: CourtDivisionDetails[] = [
+      {
+        courtDivisionCode: '001',
+        groupDesignator: 'AA',
+        courtId: '0101',
+        officeCode: '1',
+        officeName: 'A1',
+        state: 'NY',
+        courtName: 'A',
+        courtDivisionName: 'New York 1',
+        regionId: '02',
+        regionName: 'NEW YORK',
+      },
+      {
+        courtDivisionCode: '003',
+        groupDesignator: 'AC',
+        courtId: '0103',
+        officeCode: '3',
+        officeName: 'C1',
+        state: 'NY',
+        courtName: 'C',
+        courtDivisionName: 'New York 1',
+        regionId: '02',
+        regionName: 'NEW YORK',
+      },
+      {
+        courtDivisionCode: '003',
+        groupDesignator: 'AC',
+        courtId: '0103',
+        officeCode: '3',
+        officeName: 'C1',
+        state: 'NY',
+        courtName: 'C',
+        courtDivisionName: 'New York 1',
+        regionId: '02',
+        regionName: 'NEW YORK',
+      },
+      {
+        courtDivisionCode: '002',
+        groupDesignator: 'AB',
+        courtId: '0102',
+        officeCode: '2',
+        officeName: 'B1',
+        state: 'NY',
+        courtName: 'B',
+        courtDivisionName: 'New York 1',
+        regionId: '02',
+        regionName: 'NEW YORK',
+      },
+    ];
+    const expectedOffices: CourtDivisionDetails[] = [
+      {
+        courtDivisionCode: '001',
+        groupDesignator: 'AA',
+        courtId: '0101',
+        officeCode: '1',
+        officeName: 'A1',
+        state: 'NY',
+        courtName: 'A',
+        courtDivisionName: 'New York 1',
+        regionId: '02',
+        regionName: 'NEW YORK',
+      },
+      {
+        courtDivisionCode: '002',
+        groupDesignator: 'AB',
+        courtId: '0102',
+        officeCode: '2',
+        officeName: 'B1',
+        state: 'NY',
+        courtName: 'B',
+        courtDivisionName: 'New York 1',
+        regionId: '02',
+        regionName: 'NEW YORK',
+      },
+      {
+        courtDivisionCode: '003',
+        groupDesignator: 'AC',
+        courtId: '0103',
+        officeCode: '3',
+        officeName: 'C1',
+        state: 'NY',
+        courtName: 'C',
+        courtDivisionName: 'New York 1',
+        regionId: '02',
+        regionName: 'NEW YORK',
+      },
+      {
+        courtDivisionCode: '003',
+        groupDesignator: 'AC',
+        courtId: '0103',
+        officeCode: '3',
+        officeName: 'C1',
+        state: 'NY',
+        courtName: 'C',
+        courtDivisionName: 'New York 1',
+        regionId: '02',
+        regionName: 'NEW YORK',
+      },
+    ];
+    const actualOffices = testOffices.sort(courtSorter);
+    expect(actualOffices).toEqual<CourtDivisionDetails[]>(expectedOffices);
   });
 });
