+++ user-interface/src/search/SearchScreen.tsx
@@ -8,8 +8,7 @@ import { CourtDivisionDetails } from '@common/cams/courts';
 import CaseNumberInput from '@/lib/components/CaseNumberInput';
 import { useApi2 } from '@/lib/hooks/UseApi2';
 import { ComboBoxRef, InputRef } from '@/lib/type-declarations/input-fields';
-import { getOfficeList } from '@/data-verification/dataVerificationHelper';
-import { courtSorter } from '@/data-verification/DataVerificationScreen';
+import { courtSorter, getOfficeList } from '@/data-verification/dataVerificationHelper';
 import Alert, { UswdsAlertStyle } from '@/lib/components/uswds/Alert';
 import './SearchScreen.scss';
 import ComboBox, { ComboOption } from '@/lib/components/combobox/ComboBox';
